evidently_service/requirements.txt
Flask~=2.0.1
pandas~=1.3.5
Werkzeug~=2.0.1
requests~=2.26.0
python-dotenv
sqlalchemy
psycopg2-binary
pyarrow